consolidate auto-assign vpc ipv6 address logic into one component
