+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeInterfaces/AddInterfaceDrawer/VPC/AddVPCIPv6Address.tsx
@@ -1,0 +1,51 @@
+import { useSubnetQuery } from '@linode/queries';
+import { Notice, Stack } from '@linode/ui';
+import React from 'react';
+import { Controller, useFormContext } from 'react-hook-form';
+
+import { ErrorMessage } from 'src/components/ErrorMessage';
+import { VPCIPv6Address } from 'src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeInterfaces/VPCIPv6Address';
+import { generateVPCIPv6InputHelperText } from 'src/features/VPCs/utils';
+
+import type { CreateInterfaceFormValues } from '../utilities';
+
+export const AddVPCIPv6Address = () => {
+  const {
+    control,
+    getValues,
+    formState: { errors },
+  } = useFormContext<CreateInterfaceFormValues>();
+
+  const { vpc } = getValues();
+  const { data: subnet } = useSubnetQuery(
+    vpc?.vpc_id ?? -1,
+    vpc?.subnet_id ?? -1,
+    Boolean(vpc?.vpc_id && vpc?.subnet_id)
+  );
+
+  const error = errors.vpc?.ipv6?.message;
+
+  return (
+    <Stack spacing={1}>
+      {error && (
+        <Notice variant="error">
+          <ErrorMessage message={error} />
+        </Notice>
+      )}
+      <Controller
+        control={control}
+        name="vpc.ipv6.slaac.0.range"
+        render={({ field, fieldState }) => (
+          <VPCIPv6Address
+            errorMessage={fieldState.error?.message}
+            fieldValue={field.value}
+            helperText={generateVPCIPv6InputHelperText(
+              subnet?.ipv6?.[0].range ?? ''
+            )}
+            onChange={field.onChange}
+          />
+        )}
+      />
+    </Stack>
+  );
+};
